requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
chardet==5.2.0; python_version >= '3.7'
cssselect2==0.7.0; python_version >= '3.7'
django==5.0.3; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
djangorestframework==3.15.1; python_version >= '3.6'
lxml==5.2.1; python_version >= '3.6'
pillow==10.3.0; python_version >= '3.8'
python-barcode==0.15.1


sqlparse==0.4.4; python_version >= '3.5'
svglib==1.5.1; python_version >= '3.7'
tinycss2==1.2.1; python_version >= '3.7'
tzdata==2024.1; python_version >= '2'
webencodings==0.5.1
